dhall-toml: refactor modules
Broke the single Dhall.Toml module into multiple modules:
* Dhall.DhallToToml
* Dhall.TomlToDhall
* Dhall.Toml.Utils - currently has functions for parsing and normalizing
  dhall source code into expressions
