+++ dhall-toml/dhall-toml.cabal
@@ -13,12 +13,12 @@ Description:
     Use this package if you want to convert between dhall expressions and TOML.
     You can use this package as a library or an executable:
     .
-    * See "Dhall.Toml" or "Dhall.TomlToDhall" modules if you want to use this
+    * See "Dhall.DhallToToml" or "Dhall.TomlToDhall" modules if you want to use this
       package as a library
     * Use @dhall-to-toml@, @toml-to-dhall@ programs from this package if you
       want an executable.
     .
-    The "Dhall.Toml" and "Dhall.TomlToDhall" modules also contains instructions
+    The "Dhall.DhallToToml" and "Dhall.TomlToDhall" modules also contains instructions
     for how to use this package
 Category:           Compiler
 Extra-Source-Files:
@@ -41,7 +41,9 @@ Library
         containers           >= 0.5.9    && < 0.7  ,
         unordered-containers >= 0.2      && < 0.3
     Exposed-Modules:
-        Dhall.Toml
+        Dhall.DhallToToml
+        Dhall.TomlToDhall
+        Dhall.Toml.Utils
     GHC-Options:      -Wall
     Default-Language: Haskell2010
